+++ twilight-model/src/guild/role_tags.rs
@@ -1,6 +1,6 @@
 use crate::{
     id::{
-        marker::{IntegrationMarker, UserMarker},
+        marker::{IntegrationMarker, RoleSubscriptionSkuMarker, UserMarker},
         Id,
     },
     util::is_false,
@@ -12,12 +12,29 @@ use serde::{Deserialize, Serialize};
 /// [`Role`]: super::Role
 #[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
 pub struct RoleTags {
+    /// Whether this role is available for purchase.
+    #[serde(
+        default,
+        skip_serializing_if = "is_false",
+        with = "crate::visitor::null_boolean"
+    )]
+    pub available_for_purchase: bool,
     /// ID of the bot the role belongs to.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub bot_id: Option<Id<UserMarker>>,
+    /// Whether this role is a guild's linked role.
+    #[serde(
+        default,
+        skip_serializing_if = "is_false",
+        with = "crate::visitor::null_boolean"
+    )]
+    pub guild_connections: bool,
     /// ID of the integration the role belongs to.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub integration_id: Option<Id<IntegrationMarker>>,
+    /// ID of the role's subscription SKU and listing.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub subscription_listing_id: Option<Id<RoleSubscriptionSkuMarker>>,
     /// Whether this is the guild's premium subscriber role.
     #[serde(
         default,
@@ -36,9 +53,12 @@ mod tests {
     #[test]
     fn bot() {
         let tags = RoleTags {
+            available_for_purchase: false,
             bot_id: Some(Id::new(1)),
+            guild_connections: false,
             integration_id: Some(Id::new(2)),
             premium_subscriber: false,
+            subscription_listing_id: None,
         };
 
         serde_test::assert_tokens(
@@ -64,9 +84,12 @@ mod tests {
     #[test]
     fn premium_subscriber() {
         let tags = RoleTags {
+            available_for_purchase: false,
             bot_id: None,
+            guild_connections: false,
             integration_id: None,
             premium_subscriber: true,
+            subscription_listing_id: None,
         };
 
         serde_test::assert_tokens(
@@ -83,15 +106,51 @@ mod tests {
         );
     }
 
+    #[test]
+    fn subscription() {
+        let tags = RoleTags {
+            available_for_purchase: true,
+            bot_id: None,
+            guild_connections: false,
+            integration_id: Some(Id::new(1)),
+            subscription_listing_id: Some(Id::new(2)),
+            premium_subscriber: false,
+        };
+
+        serde_test::assert_tokens(
+            &tags,
+            &[
+                Token::Struct {
+                    name: "RoleTags",
+                    len: 3,
+                },
+                Token::Str("available_for_purchase"),
+                Token::None,
+                Token::Str("integration_id"),
+                Token::Some,
+                Token::NewtypeStruct { name: "Id" },
+                Token::Str("1"),
+                Token::Str("subscription_listing_id"),
+                Token::Some,
+                Token::NewtypeStruct { name: "Id" },
+                Token::Str("2"),
+                Token::StructEnd,
+            ],
+        );
+    }
+
     /// Test that if all fields are None and the optional null fields are false,
     /// then serialize back into the source payload (where all fields are not
     /// present).
     #[test]
     fn none() {
         let tags = RoleTags {
+            available_for_purchase: false,
             bot_id: None,
+            guild_connections: false,
             integration_id: None,
             premium_subscriber: false,
+            subscription_listing_id: None,
         };
 
         serde_test::assert_tokens(