+++ twilight-model/src/oauth/mod.rs
@@ -11,12 +11,14 @@ pub mod current_application_info {
 
 mod application;
 mod application_flags;
+mod current_authorization_information;
 mod install_params;
 mod partial_application;
 
 pub use self::{
-    application::Application, application_flags::ApplicationFlags, install_params::InstallParams,
-    partial_application::PartialApplication,
+    application::Application, application_flags::ApplicationFlags,
+    current_authorization_information::CurrentAuthorizationInformation,
+    install_params::InstallParams, partial_application::PartialApplication,
 };
 
 #[allow(deprecated)]