+++ twilight-http-ratelimiting/CHANGELOG.md
@@ -6,6 +6,16 @@
 
 - [**breaking**] bring widgets up to date ([#1848](https://github.com/twilight-rs/twilight/issues/1848))
 
+## [0.14.2] - 2023-01-20
+
+### Features
+
+- get current authorization route ([#2049](https://github.com/twilight-rs/twilight/issues/2049))
+
+### Refactor
+
+- remove prelude imports
+
 ## [0.14.1] - 2023-01-07
 
 ### Features